Add bit-aligned / sub-byte PDO entry support to the emulator
  Lets the emulator host slaves that declare sub-byte PDO entries (e.g.
  4 BOOL GPIOs at <BitLen>1</BitLen>):

  - EmulatedESC honors FMMU logical/physical bit offsets via a per-bit
    slow path; byte-aligned mappings keep the memcpy fast path.
    FMMUs targeting addresses < 0x1000 (e.g. mailbox-status) now work.
  - CoE::Entry gains data_bit_offset; addEntry/EsiParser allocate
    (bitlen+7)/8 bytes (was 0 for BOOL/BIT2..7). New copyBits /
    read/writeEntryBits helpers.
  - SDO upload/download/complete-access switched to bit-accurate
    payload positioning. Sub-byte entries travel as 1 octet per
    ETG1000.5 §5.3.1; CA packs entries by bitoff per ETG2000.

+++ lib/slave/include/kickcat/PDO.h
@@ -30,7 +30,7 @@ namespace kickcat
 
         std::vector<uint16_t> parseAssignment(CoE::Dictionary& dict, uint16_t assign_idx);
 
-        bool parsePdoMap(CoE::Dictionary& dict, uint16_t pdo_idx, void* buffer, uint16_t& bit_offset, uint32_t max_size);
+        bool parsePdoMap(CoE::Dictionary& dict, uint16_t pdo_idx, void* buffer, uint32_t& bit_offset, uint32_t max_size);
 
         AbstractESC* esc_;
         void* input_                = {nullptr};
